Add a fix for the configuraiton loading to resolve property mappings
Using lists of maps, the current Grails YAML processor does not correctly resolve
and spring property mappings.
This adds a fix for this, using traits to add it to all places where lists of maps
are used.
Also move the same method for configParse for property classes into a trait which
can be shared.