Migrate @Override onStartup() into Plugin File.
The **OnStartupListener** was actually being exeucted _after_ the
the `onStartup` method in an application's **Application.groovy**
class.

This had the effect of appearing to not have Consumers start up
even though `rabbitContext.startConsumers()` may have been
executed in that method; because there were no Consumers that
had been created and registered to the **RabbitContext**,
none were available to start.