Eliminate use of MimeType.getCharSet
Spring has removed the MimeType.getCharSet method entirely in favor of
the MimeType.getCharset method as of Spring 5.  Making this change
allows the rabbitmq-native plugin to function with versions of Grails
3.3 and greater, including Grails 4.

The version of the plugin has been bumped to 4.0.0 to comply with
semver, indicating a breaking change since this will mean the plugin
will not work with versions of Grails prior to 3.3.